refactor: generalize toggleExpandIcon method
