+++ src/renderer/ElectronFrame.ts
@@ -135,16 +135,6 @@ export class ElectronFrame extends Frame {
         this._setEvents()
     }
 
-    toggleExpandIcon() {
-        if (this.frameStyle === "macos") {
-            const expand_div = this.frame.querySelector("#expand") as HTMLElement
-            //Ao inserir o svg dentro de um elemento html ele muda, isso é realmente necessário para comparação
-            const temp_div = document.createElement('div')
-            temp_div.innerHTML = icons.macos.expand
-            expand_div.innerHTML = expand_div.innerHTML.trim() == temp_div.innerHTML.trim() ? icons.macos.restore : icons.macos.expand
-        }
-    }
-
     setTitle(title: string) {
         this.options.title = title
         this.update()
