+++ src/renderer/Frame.ts
@@ -1,3 +1,4 @@
+import { icons } from "./icons"
 import { format } from "./Util"
 
 export interface BaseFrameOptions {
@@ -46,6 +47,16 @@ export abstract class Frame {
 
         if (!this.frame.classList.contains("custom")) {
             this.frame.classList.add("custom")
+        }
+    }
+
+    toggleExpandIcon() {
+        if (this.frameStyle === "macos") {
+            const expand_div = this.frame.querySelector("#expand") as HTMLElement
+            //Ao inserir o svg dentro de um elemento html ele muda, isso é realmente necessário para comparação
+            const temp_div = document.createElement('div')
+            temp_div.innerHTML = icons.macos.expand
+            expand_div.innerHTML = expand_div.innerHTML.trim() == temp_div.innerHTML.trim() ? icons.macos.restore : icons.macos.expand
         }
     }
